api-gateway/requirements.txt
# Core
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
loguru==0.7.2
gunicorn==21.2.0
paho-mqtt==1.6.1

# Database
pymongo==4.6.1
redis==5.0.1

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Task Queue
celery==5.3.4

# AI/ML
openai==1.12.0
requests==2.31.0

# Analytics
pandas==2.2.0
numpy==1.26.3
streamlit==1.31.0

# Video Processing
opencv-python==4.9.0.80
moviepy==1.0.3
Pillow==10.2.0

# Utils
pydantic==2.5.3
validators==0.22.0
python-multipart==0.0.6
